Read Denovo Enriched Motifs HOMER Output

Reads in results from a denovo motif enrichment analysis created by HOMER.

read_denovo_results(path, homer_dir = TRUE)

Arguments

path
path to the HOMER directory where all outputs are stored

homer_dir
does the path point to a HOMER directory; if FALSE, path must point to the file directly [default: TRUE]

Value

a tibble with the following columns:

consensus the consensus sequence of the denovo motif

motif_name name of the motif

log_odds_detection threshold used to determine bound vs. unbound sites

motif_pwm a list column with PWMs for each motif

log_p_value_detection from the original experiment used to ID motif

tgt_num number of times motif appears in target sequences

tgt_pct percent of times motif appears in target sequences

bgd_num number of times motif appears in background sequences

bgd_pct percent of times motif appears in background sequences

log_p_value final enrichment from experiment -log10(p-value)

tgt_pos average position of motif in target sequences, where 0 = start of sequences

tgt_std standard deviation of position in target sequences

bgd_pos average position of motif in background sequences, where 0 = start of sequences

bgd_std standard deviation of position in background sequences

strand_bias log ratio of + strand occurrences to - strand occurrences

multiplicity average number of occurrences per sequence in sequences with 1 or more binding sites

Details

Following an analysis using find_motifs_genome, a HOMER directory is created which analyses the enrichment of denovo motifs. This function reads and parses in that file into a tidy format.

See also